--- sqlglot/engine/__init__.py
@@ -1,1 +1,0 @@
-from sqlglot.engine.engine import execute
